pyproject.toml
[tool.poetry]
name = "tds-proj2"
version = "0.1.0"
description = "A project to explore data science concepts and techniques using real-world datasets."
authors = ["Alakhya Sarkar <23f1001050@ds.study.iitm.ac.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13"
alembic = ">=1.14.0"
fastapi = ">=0.115.6"
sqlalchemy = ">=2.0.36"
uvicorn = ">=0.32.1"
pandas = "^2.2.3"
seaborn = "^0.13.2"
matplotlib = "^3.9.4"
openai = "0.28"
ipykernel = "^6.29.5"
scikit-learn = "^1.6.0"
httpx = "^0.28.1"
chardet = "^5.2.0"
scipy = "^1.14.1"
python-dotenv = "^1.0.1"
setuptools = "^75.6.0"
wheel = "^0.45.1"

[tool.poetry.dev-dependencies]
ruff = ">=0.8.2"

[project]
name = "tds-proj2"
version = "0.1.0"
description = "A project to explore data science concepts and techniques using real-world datasets."
authors = ["Alakhya Sarkar <23f1001050@ds.study.iitm.ac.in>"]
dependencies = [
    "pandas",
    "seaborn",
    "matplotlib",
    "openai==0.28"
]
